Port tests to VMx
The tests are equivalent to the ones in "hatchery" but propose a
different API for interacting with the virtual machine.

The main entrypoint is the `VM` struct. Creating an instance using `new`
or `load` - ephemeral and persistent versions respectively. One can
then request a session from a `VM` using the `session` or `session_mut`
functions, which respectively return a `Session` and a `SessionMut`
which allow to perform `query`es and `transact`ions - in the case of
`SessionMut`.

`SessionMut` instances can be persisted using the `VM::commit` function,
returning a `CommitId`, which can then be used to instantiate sessions
starting from the given persisted state.