Implement module metering
To make this possible, a running instance has to be made accessible
through an `Env`. In this implementation, we purposefully leak the
instance and place a pointer to it in the call stack. While the stack
element exists in the call stack, mutable references to this instance
are valid. When this element is dropped, all references to it become
invalid.

This works due to the fact that the stack element is guaranteed to
exists for the duration of a call.

Resolves #83