Enforce a single instance per module
Ensures a module is only ever instantiated once per session, by using a
map in the call stack. If an instance already exists in an instance map,
then an element is appended without creating a new instance. If an instance
does not exist in the map, a new one is created and then a new element
appended to the stack.

Resolves #122