Implement call atomicity
Each call, be it a transaction or a query, is now atomic - meaning it
will not change the state if it fails. This is also true for every
inter-contract call performed during said call.

This is achieved by a re-execution strategy. Since we have no way of
partially rolling back changes in memory, each failing call - and
inter-contract for that matter - will trigger a re-execution of all
historic calls in a session until said failure occurs. The failure is
then skipped, communicated to the caller, and execution proceeds until
the next error - if so the procedure is gone through again - or until it
succeeds.

Resolves #155
Resolves #156