Fix dangling pointer after session move
This fix introduces a `SessionInner`, containing the previous contents
of the `Session`, and changes the fields of a session to contain a
`&'static mut SessionInner` and a reference count.

We can then internally clone a session, which aliases the reference to
the `SessionInner`, allowing us to place a session directly in the
`Env`.

It is then ensured that a session frees the `SessionInner` on `Drop`ping
when the reference count reaches zero.

Resolves: #202