uplink: add more comprehensive documentation
We add some crate level docs, ensuring the first page of `docs.rs` is
not so barebones, and ensure all relevant functions have at least a
small comment adorning them.

To be more explicit on the feature requirements of each function, we
also add some compile-time shennanigans for us to detect when `docs.rs`
is building our crate, to properly display the required features on each
function and struct.

Resolves: #190
Resolves: #189
See-also: #195