piecrust: port to use `crumbles` for memories
This makes it possible for us to remove diffing, in favor of iterating
through each page, and only writing them one dirty page at a time during
each commit. One interesting consequence is that squashing commits no
longer makes sense., and is thus removed from the code.

Each page is part of a merkle tree, making it possible to compute a
*proof of inclusion for each page* and present it to some user. It might
be interesting to expose this in the future.