piecrust: fix reporting of memory to host
The reporting of memory to the host was previously based on the growth
offset reported to the contract, whereas, as far as both the host and
the contract are concerned, the entire memory is already allocated and
available for writing by both parties. This is fixed by acknowledging
this on the host end, and slicing the memory appropriately.

Some bandaids to "fix" this behavior are removed, such as fake guard
pages, and the growth behavior of memory is comprehensive now.

See-also: https://github.com/dusk-network/rusk/issues/1033