crumbles: allow for specifying mapping size
This requires changing the constructors, to be able to support the
change. We choose having the user specify the size of a page and their
number, and computing the length of the memory internally. To ensure
that the dirt-tracking mechanism stays intact, the page size must be a
multiple of the system page size.