crumbles: add lazy page mappings
This is achieved by having the caller of `Mmap::with_files` provide an
implementation of the `LocateFile` trait, which is called when a page is
first being accessed to return the (optional) path to the file that is
supposed to be mapped to the given page index.

Due to the mechanism required to detect any access (mprotect), two bits
per page need to be kept. These are one bit for discerning if a file has
been mapped to that page, and another for testing whether it has been
accessed (read/write) once. This mechanism allows for marking each page
dirty, using the same rules as before.