Use Dusk compiler for contracts
The compiler is downloaded and extracted to the target directory, and
linked to the toolchains in the rustup directory, making it available to
be used anyone in the system.

The compiler is available as under the `+dusk` option in rust tools, and
is able to compile at least the `wasm32-unknown-unknown` target.

See-also: #280