uplink: remove `RawCall` and `RawResult`
It can be noticed that these types in reality are unnecessary for the
library in that they can easily be subtituted by using the `String` and
`Vec` types in the standard library.

This also leads to a less confusing call handling for external users,
and clears the path for the implementation of rich panic information.

See-also: #301