piecrust: change `owner` to accept any contract ID
The `owner` import is changed to accept a pointer to a contract ID. If
the pointer is null, it assumed that the calling contract is querying
for its owner, whereas if the pointer is not null, a contract ID is
presumed to exist at that address, and read from its memory.

The relevant owner is then emplaced in the contract's argument buffer,
if it exists, and 1 is returned. If the contract whose owner is asked
for does not exist, 0 is returned instead.

Co-authored-by: Milosz Muszynski <milosz@dusk.network>