contracts: add `double_counter` contract
This contract is intended to be a near copy of the `counter` contract,
but containing two different integers as opposed to just one. It will be
used to test migrating from the counter contract to this new contract -
i.e. proof that it is possible to substitute contracts in the state.