piecrust: use empty constructor arguments by default
Previously we were erroring out when providing no arguments for a
constructor that takes no argument. This is undesirable since downstream
does not necessarily know that the contract they deploy has a constructor.

A contract that is provided an argument that it doesn't expect will
still error at deploy time, during execution of the constructor.

Resolves #316