piecrust: change `caller` import to return integer
The integer signals whether there is actually a caller or not. It will
be 1 if there is a caller, 0 if no caller, and the contract is the first
in the call chain.

This allows for contracts to more easily handle the situation in which
they're the first to be called, a situation which happens often
downstream in e.g. queries.